requirements.txt
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-openai==1.1.14
langchain-text-splitters==1.1.2
faiss-cpu==1.13.2
numpy==2.4.4
openai==2.32.0
tiktoken==0.12.0
